Redesign AsymmetricGridViewAdapter.
Turns AsymmetricGridViewAdapter into a wrapper that listens to changes
on the actual listView adapter and reacts to it, intead of holding its
own copy of the data. This allow more flexibility on user interaction,
allowing a CursorAdapter to be used for example. The demo app has been
updated accordingly to reflect the changes.