add raison d'etre. Modified key class constructors
This repository has a new feature. Instead of explicitly wrapping your adapter in the AsymmetricGridViewAdapter you simply call AsymmetricGridView.setAdapter( adapter ) with any old adapter and then it is wrapped and linked under the hood. This constructors are also more consistent with what an Android developer might expect. You no longer need to pass the listView as the listView will do this itself when you call setAdapter. sneaky.